requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
huggingface-hub>=0.20.0
timm>=1.0.0
streamlit>=1.30.0
grad-cam>=1.5.0
matplotlib>=3.8.0
opencv-python>=4.9.0
google-generativeai>=0.8.0
numpy>=1.23.0
pandas>=1.5.0
pytorch-grad-cam>=1.4.0
